Added wheel colliders
Wheel colliders are what we should use to move the car.  However wheel
colliders have to be separate objects from the wheel models themselves
so a new group of wheel colliders was made.  When wheel colliders move
the car using the motor torque property, the wheel models should rotate
as the animation.