AI update/overhaul
AI no works on triggers.  Trigger is on an empty object and looks for
parts labeled as the opposing force(ally v enemy).  Then the target is
set, and the tank's sighted boolean is set to true.

It currently can tell when there are multiple enemies within it's sight,
but does not keep track of more than the first one.  We might want to
put some sort of smart targeting stuff in the trigger script's update
function.  The AI also does not retarget if it's target leaves it's
sight and there is another enemy inside that radius.