AI Movement Controller works now
We can tell the tanks to move forward, turn left, turn right and go
backwards now.  The methods will do one of the actions above for a
certain time.  The stop time is assigned based on your input in the
function and the update function will check if Time.time is greater
than the stopTime and if it is then the wheels will stop.